Timeout preemption:
When a number of services has already restarted such that even if every
remaining service were to time out, we return early from the goroutines
observing all the remaining in-flight restarts, treating them very much
as if they had timed out. The deploy is still considered a success, but
we don't stick around to watch the last batch of restarts.

It is still possible that a later restart we now fail to observe *fails*
somehow, but it is rare enough in practice (i.e. I've never seen it)
that I'm ok making that tradeoff for faster deploys.